feat: new UI context and modify related components and tests
