+++ packages/frontend/src/layout/banner/banner.tsx
@@ -2,11 +2,14 @@ import { useEffect, useState, useContext } from 'react'
 import { observer } from 'mobx-react-lite'
 
 import UserContext from '../../context/User'
+import UIContext from '../../context/UI'
 
 import { ABOUT_URL } from '../../config'
 
 const Banner = () => {
     const userContext = useContext(UserContext)
+    const uiContext = useContext(UIContext)
+
     const [on, setOn] = useState<boolean>(false)
 
     useEffect(() => {
@@ -17,12 +20,12 @@ const Banner = () => {
 
     const closeBanner = () => {
         setOn(false)
-        userContext.setHasBanner(false)
+        uiContext.setHasBanner(false)
     }
 
     return (
         <div className="banner-row">
-            {userContext.hasBanner && on && (
+            {uiContext.hasBanner && on && (
                 <div className="banner">
                     <img src={require('../../../public/images/banner.svg')} />
                     <div className="banner-title">
